reduces physics runtime for large graphs

+++ nakar-server/src/lib/room/RoomService.ts
@@ -816,7 +816,7 @@ export class RoomService implements ApplicationService {
 
     this._sendActionToWorker(params.roomId, {
       type: 'WTActionTriggerPhysics',
-      amount: 'long',
+      amount: 'short',
     });
   }
